Add Robin Weiruch Tutorial, except TypeScriptified!
This tutorial was neat. I did it at night.
I eventually got tired of making interfaces and finished the tutorial
with noImplicitAny=false so I could watch tv with my girlfriend.

But I see why that is not ideal. I really look forward to knowing
types constantly on the front end and not having to google the members
of objects over and over, just having the autorecommended to me.